pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ardupyrpc"
version = "0.1.0"
description = "Python client for an Arduino RPC bridge using MsgPack over serial."
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name = "Luca Macavero" },
]
license = "MIT"
keywords = ["arduino", "teensyduino", "rpc", "serial", "msgpack"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

dependencies = [
  "pyserial",
  "msgpack",
]

[project.urls]
"Homepage" = "https://github.com/lu-maca/pyrpc"

[tool.setuptools.packages.find]
where = ["."]
